amd_ucode_info: support concatenated containers
Several containers can be and are concatenated (for example,
in the early cpio), as mentioned, for example, in [1].  Moreover,
in-kernel microcode parser supports concatenated containers as well.

[1] https://xenbits.xen.org/docs/unstable/misc/amd-ucode-container.txt

* amd_ucode_info.py (parse_equiv_table): Add start_offset argument, add
the offset in the table_item and table_stop calculations.
(parse_ucode_file): Add start_offset;  append information about
the offset to the "Microcode patches" header if it is non-zero, add
start_offset to ucode_file.seek() statements with absolute values
and to cursor calculation;  pass start_offset to the parse_equiv_table
call;  read the patch type bytes first, compare them against the magic,
return the offset if it matches, and perform the normal patch type check
otherwise;  return on the incorrect patch type rather than break the
loop.
(parse_ucode_file): Call parse_ucode_file in a loop as long as it
returns an offset to restart parsing from.

Signed-off-by: Eugene Syromiatnikov <esyr@redhat.com>