amd_ucode_info: support multiple CPUIDs for a single equiv_id
microcode_amd.bin contains several records with the same equiv_id in its
equivalence table (for example, 10-02-02/10-02-03,
or 10-04-03/10-05-03/10-06-03), and not showing the relevant CPUIDs
in the output can be considered misleading.

* amd_ucode_info.py (parse_equiv_table): Store CPUIDs associated
with each specific equiv_id in an array.
(parse_ucode_file): Iterate over the array and print the patch
information line for each CPUID.

Signed-off-by: Eugene Syromiatnikov <esyr@redhat.com>