amd_ucode_info: do not skip patches with equiv_id not in the equivalence table
While the patches without the associated CPUIDs are useless from the
kernel's point of view, not reporting their presence might be misleading.
This also eliminates an additional occurrence of the same cursor update
logic, leaving only one place where it is updated.

* amd_ucode_info.py (parse_ucode_file): Do not update cursor
and do not continue to the next patch header if there is no equiv_id
in the ids table, print some stub information instead.

Signed-off-by: Eugene Syromiatnikov <esyr@redhat.com>