amd_ucode_info: convert extract_patch to operate on patch entries
This is needed for the following patches.

* amd_ucode_info.py (PatchEntry): New namedtuple constructor.
(extract_patch): Accept out_dir;  use pass patch_start, patch_length,
and ucode_level in the patch argument;  update the implementation
and docstring accordingly.
(parse_ucode_file): Create a patch entry containing identifying
information about the patch and pass it to the extract_patch call, along
with opts.extract as the output directory.

Signed-off-by: Eugene Syromiatnikov <esyr@redhat.com>