amd_ucode_info: add an option to split out container
Unfortunately, as it is now, the files produced with the -e option
are not only unusable by the kernel as is, but also lose the information
about the CPUIDs they are associated with, which complicates any further
possible microcode manipulations.  As a possible solution, provide
an alternative solution for getting individual patches, that stores them
alongside the container header and the relevant equivalence table
entries.

* amd_ucode_info.py (EQ_TABLE_TYPE, PATCH_TYPE): New named constants.
(extract_patch): Add equiv_table argument, that defaults to None;
if equiv_table is not None, the file name pattern is changed to include
equiv_id and the associated CPUIDs;  write out the container header,
equivalence table header, and the relevant equivalence table entries
(the ones with the same equiv_id as the patch has) before writing out
the patch if equiv_table is not None;  update the docstring.
(parse_ucode_file): Change patch_type comparison vale from 1 to the
PATCH_TYPE named constant;  call extract_patch with the relevant options
(including the ids table) if opts.split is set.
(parse_options): Add -s/--split option.
* README.md: Update the "./amd_ucode_info.py --help" output;  mention
the ability to split out the container now, along with the usage
example.

Signed-off-by: Eugene Syromiatnikov <esyr@redhat.com>