amd_ucode_info: split out microcode write function
Factor out microcode data writing routine into a separate function,
leaving out the decisions on the file name and the equivalence table
contents, which is needed for the next patch.

* amd_ucode_info.py (write_mc): New function, renamed from extract_patch
and implements only writing of the equivalence table and the patches
to the supplied file object.  Uses the supplied ucode_file as the input
file if it is not None, or opens it itself otherwise.
(extract_patch): Convert to a wrapper function that decides on the output
file name and opens it, as well as on the set of equivalence table entries
to write out, then pass the file object and the equivalence table
to the write_mc function.

Signed-off-by: Eugene Syromiatnikov <esyr@redhat.com>