amd_ucode_info: exit with non-zero code on error
To be able to catch the presence of issues without the need to inspect
the output.

* amd_ucode_info.py: Import errno module.
(parse_ucode_file): Return the error code in the tuple.
(parse_ucode_files): Capture the error code returned by
parse_ucode_file, store it in the status variable, return status.
(parse_options): Return errno.ENOENT if a path passed
in the conainer_file argument is not a regular file.
(main): Exit with the status code returned by the parse_ucode_files
call.

Signed-off-by: Eugene Syromiatnikov <esyr@redhat.com>